+++ scripts/cluster.js
@@ -173,6 +173,10 @@ const commonOptions = require('./commonOptions');
                     loggerOptions.fileName = options.output;
                 }
 
+                if (options.errorFile) {
+                    loggerOptions.errorFile = options.errorFile;
+                }
+
                 logger = Logger.getLogger(loggerOptions);
                 ipc.setLoggerOptions(loggerOptions);
                 util.setLoggerOptions(loggerOptions);
@@ -189,14 +193,22 @@ const commonOptions = require('./commonOptions');
 
                 for (let i = 0; i < REQUIRED_OPTIONS.length; i++) {
                     if (!options[REQUIRED_OPTIONS[i]]) {
-                        logger.error(REQUIRED_OPTIONS[i], 'is a required command line option.');
-                        return;
+                        const error = `${REQUIRED_OPTIONS[i]} is a required command line option.`;
+
+                        ipc.send(signals.CLOUD_LIBS_ERROR);
+
+                        util.logError(error, loggerOptions);
+                        util.logAndExit(error, 'error', 1);
                     }
                 }
 
                 if (!options.password && !options.passwordUrl) {
-                    logger.error('One of --password or --password-url is required.');
-                    return;
+                    const error = 'One of --password or --password-url is required.';
+
+                    ipc.send(signals.CLOUD_LIBS_ERROR);
+
+                    util.logError(error, loggerOptions);
+                    util.logAndExit(error, 'error', 1);
                 }
 
                 // When running in cloud init, we need to exit so that cloud init can complete and
@@ -391,7 +403,12 @@ const commonOptions = require('./commonOptions');
                             }
                         }
 
-                        util.logAndExit(`Cluster failed: ${message}`, 'error', 1);
+                        ipc.send(signals.CLOUD_LIBS_ERROR);
+
+                        const error = `Cluster failed: ${message}`;
+                        util.logError(error, loggerOptions);
+                        util.logAndExit(error, 'error', 1);
+
                         exiting = true;
                         return q();
                     })
@@ -400,14 +417,15 @@ const commonOptions = require('./commonOptions');
 
                         if (!rebooting) {
                             util.deleteArgs(ARGS_FILE_ID);
-                            ipc.send(options.signal || signals.CLUSTER_DONE);
+
                             if (!exiting) {
+                                ipc.send(options.signal || signals.CLUSTER_DONE);
                                 util.logAndExit('Cluster finished.');
                             }
                         } else if (!options.reboot) {
                             // If we are rebooting, but we were called with --no-reboot, send signal
-                            ipc.send(options.signal || signals.CLUSTER_DONE);
                             if (!exiting) {
+                                ipc.send(options.signal || signals.CLUSTER_DONE);
                                 util.logAndExit('Cluster finished. Reboot required but not rebooting.');
                             }
                         } else {
@@ -417,6 +435,13 @@ const commonOptions = require('./commonOptions');
                         if (cb) {
                             cb();
                         }
+                    });
+
+                // If another script has signaled an error, exit, marking ourselves as DONE
+                ipc.once(signals.CLOUD_LIBS_ERROR)
+                    .then(() => {
+                        ipc.send(options.signal || signals.CLUSTER_DONE);
+                        util.logAndExit('ERROR signaled from other script. Exiting');
                     });
 
                 // If we reboot due to some other script, exit - otherwise cloud providers